if dividing by zero, instead set scaling factor to 0.
necessary in cases were an added variable is 0 for one year, but the market is still in use. for example no high votlage pwoer line sin 1890. It is tricky to divide markets into high/medium/low voltage (which would deal with the problem of zero division) because some of the used technologies change voltage over time